refactor: 교육 관련 서비스 간 직접 의존성 제거 및 중계 서비스 추가
## 주요 내용
`education`, `educationEnrollment`, `educationTerm`, `educationSession`, `sessionAttendance` 관련 서비스 간의 직접적인 의존성을 제거하고,
각 서비스를 연결하는 중계 서비스를 추가하여 관리하도록 구조를 변경하였습니다.

## 세부 내용
- 개별 서비스 간 직접적인 의존성 제거
- 각 두 개의 서비스를 연결하는 중계 서비스 추가
- 서비스 간 결합도를 낮추고, 유지보수성을 향상시키기 위한 구조 개선
- 기존 로직에서 중계 서비스를 활용하도록 수정

이번 변경을 통해 서비스 간 역할이 보다 명확해졌으며, 모듈 간 결합도를 줄여 확장성과 유지보수성을 향상시켰습니다. 🚀
