+++ backend/src/churches/management/const/education/education-event.const.ts
@@ -1,0 +1,3 @@
+export const EducationEventConst = {
+  educationNameUpdate: 'education-name.updated',
+};
