+++ backend/src/churches/management/event/education-name-update.event.ts
@@ -1,0 +1,9 @@
+export class EducationNameUpdateEvent {
+  constructor(
+    public readonly educationId: number,
+    public readonly educationName: string,
+
+    public readonly attempt: number = 1,
+    public readonly maxAttempts: number = 5,
+  ) {}
+}
